pyproject.toml
[project]
name = "krayne"
version = "0.2.0"
description = "CLI and SDK for creating, managing, and scaling Ray clusters on Kubernetes"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
keywords = ["ray", "kubernetes", "kuberay", "ml", "distributed-computing", "cluster-management", "cli", "sdk"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: System :: Clustering",
    "Topic :: System :: Systems Administration",
    "Typing :: Typed",
]
dependencies = [
    "typer>=0.9,<1",
    "rich>=13,<14",
    "pydantic>=2,<3",
    "kubernetes>=28,<32",
    "pyyaml>=6,<7",
    "questionary>=2,<3",
    "textual>=0.75,<1",
]

[project.urls]
Homepage = "https://github.com/roulbac/krayne"
Documentation = "https://roulbac.github.io/krayne/"
Repository = "https://github.com/roulbac/krayne"
Issues = "https://github.com/roulbac/krayne/issues"
Changelog = "https://github.com/roulbac/krayne/releases"

[project.scripts]
krayne = "krayne.cli.app:app"
ikrayne = "krayne.tui.entry:main"

[dependency-groups]
dev = [
    "pytest>=7",
    "pytest-asyncio>=0.23",
    "pytest-timeout>=2",
    "mkdocs>=1.6",
    "mkdocs-material[imaging]>=9.5",
]

[tool.uv]
default-groups = "all"

[build-system]
requires = ["uv_build>=0.11.1,<0.12"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "krayne"

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
]
